+++ tests/files/src/0870_noop_switch.php
@@ -1,0 +1,9 @@
+<?php
+function test_switch($x) {
+    switch($x) {
+    }
+    switch($x) {
+    default:
+    echo "In default\n";
+    }
+}